Redefine I3PropagatorService interface to avoid quadtratic behavior in high-multiplicity events. See #1034.
Previously, each propagator received a pointer to the current frame along
with every particle to propagate, and was allowed to fill ancillary
information (e.g. I3MMCTrackList or I3MCWeightDict) directly. Since
I3Frame is write-once, however, I3PropagatorServicePROPOSAL had to make
a copy of the existing MMCTrackList every time it wanted to append a new
entry, making the total running time quadtratic in the number of muons in
the event. To fix this without adding undue const_cast evil, we replace
I3Frame with I3PropagatorService::DiagnosticMap, a read-write analogue of
I3Frame. Each propagator can place objects in the map, but is also allowed
to modify them in subsequent calls. At the end of the propagation loop,
I3PropagatorModule transfers the objects in the map to the frame. This
keeps the interface stateless without sacrificing flexibility.

This reduces the time to propagate 100k vertical muons from 45 to
6 seconds.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/sim-services/trunk@134491 16731396-06f5-0310-8873-f7f720988828